Enum ra_project_model::ProjectWorkspace

pub enum ProjectWorkspace {
    Cargo {
        cargo: CargoWorkspace,
        sysroot: Sysroot,
    },
    Json {
        project: JsonProject,
    },
}

Variants

Cargo

Project workspace was discovered by running cargo metadata and rustc --print sysroot.

Fields of Cargo

cargo: CargoWorkspacesysroot: Sysroot

Json

Project workspace was manually specified using a rust-project.json file.

Fields of Json

project: JsonProject

Methods

impl ProjectWorkspace

pub fn load(
root: ProjectRoot,
cargo_features: &CargoConfig,
with_sysroot: bool
) -> Result<ProjectWorkspace>

pub fn to_roots(&self) -> Vec<PackageRoot>

Returns the roots for the current ProjectWorkspace The return type contains the path and whether or not the root is a member of the current workspace

pub fn out_dirs(&self) -> Vec<PathBuf>

pub fn proc_macro_dylib_paths(&self) -> Vec<PathBuf>

pub fn n_packages(&self) -> usize

pub fn to_crate_graph(
&self,
default_cfg_options: &CfgOptions,
extern_source_roots: &FxHashMap<PathBuf, ExternSourceId>,
proc_macro_client: &ProcMacroClient,
load: &mut dyn FnMut(&Path) -> Option<FileId>
) -> CrateGraph

pub fn workspace_root_for(&self, path: &Path) -> Option<&Path>